requirements.txt
config==0.5.1
DateTime==5.5
fernet==1.0.1
firebase-admin==6.5.0
Flask==3.0.3
folium==0.12.0
gcloud==0.18.3
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
py==1.11.0
python-dotenv==1.0.1
gunicorn==23.0.0